# server-go/api/events.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
)

// Event defines model for Event.
type Event struct {
	// Date The scheduled date and time of the event
	Date time.Time `json:"date"`
	Id   string    `json:"id"`

	// Location Information about the venue where the event will be held
	Location Location `json:"location"`

	// Name The name of the event
	Name string `json:"name"`
}

// EventError Error
type EventError struct {
	Code    int32  `json:"code"`
	Message string `json:"message"`
}

// Location defines model for Location.
type Location struct {
	Name string `json:"name"`
}

// EventsFindByNameFuzzyParams defines parameters for EventsFindByNameFuzzy.
type EventsFindByNameFuzzyParams struct {
	Name string `form:"name" json:"name"`
}

// EventsCreateJSONRequestBody defines body for EventsCreate for application/json ContentType.
type EventsCreateJSONRequestBody = Event

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /events)
	EventsFindByNameFuzzy(c *gin.Context, params EventsFindByNameFuzzyParams)

	// (POST /events)
	EventsCreate(c *gin.Context)

	// (DELETE /events/{id})
	EventsDelete(c *gin.Context, id string)

	// (GET /events/{id})
	EventsRead(c *gin.Context, id string)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// EventsFindByNameFuzzy operation middleware
func (siw *ServerInterfaceWrapper) EventsFindByNameFuzzy(c *gin.Context) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params EventsFindByNameFuzzyParams

	// ------------- Required query parameter "name" -------------

	if paramValue := c.Query("name"); paramValue != "" {

	} else {
		siw.ErrorHandler(c, fmt.Errorf("Query argument name is required, but not found"), http.StatusBadRequest)
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "name", c.Request.URL.Query(), &params.Name)
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter name: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.EventsFindByNameFuzzy(c, params)
}

// EventsCreate operation middleware
func (siw *ServerInterfaceWrapper) EventsCreate(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.EventsCreate(c)
}

// EventsDelete operation middleware
func (siw *ServerInterfaceWrapper) EventsDelete(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.EventsDelete(c, id)
}

// EventsRead operation middleware
func (siw *ServerInterfaceWrapper) EventsRead(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id string

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.EventsRead(c, id)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.GET(options.BaseURL+"/events", wrapper.EventsFindByNameFuzzy)
	router.POST(options.BaseURL+"/events", wrapper.EventsCreate)
	router.DELETE(options.BaseURL+"/events/:id", wrapper.EventsDelete)
	router.GET(options.BaseURL+"/events/:id", wrapper.EventsRead)
}